chore: refactor the rest of the modals, remove <modal> component
