+++ packages/autopilot/src/app/components/index.ts
@@ -24,7 +24,6 @@ import InsertLine from './insert-line.vue';
 import LinkAction from './link-action.vue';
 import LinkContext from './link-context.vue';
 import Loader from './loader.vue';
-import Modal from './modal.vue';
 import PromoRobotSchool from './promo-robot-school.vue';
 import Sb from './sb.vue';
 import SigninWarning from './signin-warning.vue';
@@ -41,7 +40,6 @@ Vue.component('insert-line', InsertLine);
 Vue.component('link-action', LinkAction);
 Vue.component('link-context', LinkContext);
 Vue.component('loader', Loader);
-Vue.component('modal', Modal);
 Vue.component('promo-robot-school', PromoRobotSchool);
 Vue.component('sb', Sb);
 Vue.component('signin-warning', SigninWarning);
